Access of multiple user according to their role @PreAuthorize anotation

+++ src/main/java/com/springSecurity/SpringSecurity/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package com.springSecurity.SpringSecurity.repository;
 
-import com.springSecurity.SpringSecurity.model.User;
+import com.springSecurity.SpringSecurity.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User,Short> {
+public interface UserRepository extends JpaRepository<User,Long> {
 }
